#167 [move] : change directory
