+++ core/ui/src/main/res/navigation/graph_ui.xml
@@ -29,7 +29,7 @@
     </dialog>
     <dialog
         android:id="@+id/navigation_feed_image_dialog"
-        android:name="com.teamwable.ui.component.FeedImageDialog"
+        android:name="com.teamwable.ui.component.feedimage.FeedImageDialog"
         android:label="FeedImageDialog"
         tools:layout="@layout/dialog_feed_image">
         <argument
